Update AddressForm to have the Checkout button in it, Delete the Checkout button component
